Update README.md
The system consists of an admin module where the admin can create questions. The marks for the correct answers and negative marks for wrong answers are set in the creation of the questions. This system has an aim to set the internal assessment of the lab in the college.
The admin module creates questions. The admin has a permanent email address and password stored in the database. So, admin can log in using this id and password. Students can read the questions and answer to it. Student module have a section where the users can report complaints about the computer. Those complaints can be viewed by the admin. New student can register with their name, email, and password. This system is a simple web application in which we have an exam and quiz management. Admin gives the permission to enter to this system. Students can attach files to assignment section.